% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod.R
\name{prod}
\alias{prod}
\title{Product}
\usage{
prod(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
Returns the product of \code{x} and \code{y}.
}
\description{
\code{product} calculate the product of two numbers
}
\examples{
# Generate example data
set.seed(123)
x <- 5
y <- 4
z <- 5
# Compute MDC with U-centering
product <- prod(x,y)


}
